LayoutTests/imported/w3c/web-platform-tests/xhr/setrequestheader-header-forbidden-expected.txt

CONSOLE MESSAGE: line 16: Refused to set unsafe header "Accept-Charset"
CONSOLE MESSAGE: line 17: Refused to set unsafe header "Accept-Encoding"
CONSOLE MESSAGE: line 18: Refused to set unsafe header "Connection"
CONSOLE MESSAGE: line 19: Refused to set unsafe header "Content-Length"
CONSOLE MESSAGE: line 20: Refused to set unsafe header "Cookie"
CONSOLE MESSAGE: line 21: Refused to set unsafe header "Cookie2"
CONSOLE MESSAGE: line 22: Refused to set unsafe header "Date"
CONSOLE MESSAGE: line 23: Refused to set unsafe header "DNT"
CONSOLE MESSAGE: line 24: Refused to set unsafe header "Expect"
CONSOLE MESSAGE: line 25: Refused to set unsafe header "Host"
CONSOLE MESSAGE: line 26: Refused to set unsafe header "Keep-Alive"
CONSOLE MESSAGE: line 27: Refused to set unsafe header "Referer"
CONSOLE MESSAGE: line 28: Refused to set unsafe header "TE"
CONSOLE MESSAGE: line 29: Refused to set unsafe header "Trailer"
CONSOLE MESSAGE: line 30: Refused to set unsafe header "Transfer-Encoding"
CONSOLE MESSAGE: line 31: Refused to set unsafe header "Upgrade"
CONSOLE MESSAGE: line 32: Refused to set unsafe header "Via"
CONSOLE MESSAGE: line 33: Refused to set unsafe header "Proxy-"
CONSOLE MESSAGE: line 34: Refused to set unsafe header "Proxy-LIES"
CONSOLE MESSAGE: line 35: Refused to set unsafe header "Proxy-Authorization"
CONSOLE MESSAGE: line 36: Refused to set unsafe header "Sec-"
CONSOLE MESSAGE: line 37: Refused to set unsafe header "Sec-X"

PASS XMLHttpRequest: setRequestHeader() - headers that are forbidden 
